4.21: blogs can only be deleted by user that created it

+++ part4/bloglist/backend/controllers/blogs.js
@@ -55,8 +55,27 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const body = request.body
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if(!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    if(!user) {
+        return response.status(400).json({ error: 'userId missing or not valid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if(blog.user.toString() === user._id.toString()) {
+        await Blog.findByIdAndDelete(request.params.id)
+        return response.status(204).end()
+    }
+    return response.status(400).json({ error: 'blog to delete is not created by this user' })
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
